Zooming with mouse wheel centered to cursor
This happens except when the image is smaller than the window, in which
case the constraint in the position is to center the image.

The system of coordinates has been changed to using a QTransform for the
zoom instead of manually stretching or squeezing the coordinates of
inner and outer rectangles, points center and correlation maps.